Properly indented the code
CTRL-A to select all, CTRL-ALT-I to indent. Got it Alex? Also reduced
closing tags that didn't close anything. Added a log out button that
brings the user to the intro page. What needs to be fixed is your grid
system code that contains nothing with your content being nested outside
of it.